requirements.txt
boto3==1.36.0
Flask==3.1.0
Flask-WTF==1.2.2
pillow==11.1.0
